Ditched the lexer.
Unfortunately the megaparsec library makes it a bit difficult to
handle a separate lexer while keeping track of current position (the
getSourcePos operation is supposedly slow acording to the documentation
(i suspect it goes through the entire file to calculate line and column
numbers based on the stored offset) and it would have to be called for
every token).

Since getSourcePos 'is not cheap' I've introduced another way to store
positions in AST. Instead of extracting the SourcePos from the parser
state, I only extract the file name and the offset, which is cheap.
Then, if there is an error we can simply take the offset and find the line
and column numbers ourselves.
The AST types have SourcePos baked-in unfortunately so we will need
to change it in a future pull request.