'Utility' module.
Some things will have to be shared across the entire codebase, so
I created a separate module, which will contain things like
useful functions that don't fit another modules well (or would be
imported from many places).

One of such things is the new 'Position' type that will replace
'SourcePos' from Megaparsec. While SourcePos is a convenient
way to store positions of things, it is unfortunately difficult
to extract while parsing (in a efficient way). Instead we will
simply store the file name and the offset from the beginning. It
will be sufficient to print nice error messages (line and column
numbers can be easily recovered using the offset) and is much
easier to extract from the parser state.

I have replaced all SourcePos occurences in the Syntax.hs with Position.