Defined the Parser monad.
Since our grammar is tricky to parse, we needed the following features:
- reading new files during parsing
- passing and updating a dictionary of custom operators

To achieve that, I used monad transformers to create the parser type
that is able to both perfom stateful computations and side effects.

The Parser type wraps the IOState type using ParsecT monad transformer.
The IOState is the IO monad wrapped inside StateT transformer.