Enchanced Parser with ReaderT.
I decided it would be convenient to be able to access the
command line options during parsing without explicitly passing
it around as an argument. To do that, I added yet another
monad transformer to the stack.
Now the stack looks like this:
	ParsecT <- StateT <- ReaderT <- IO
Reader monad allows the retrieval of a shared environment.