Change in let AST.
Previously we had two types of patterns that were possible in a
let definition:
- LetPattern (Pattern a)
- FuncPattern Identifier [Pattern a]
They were defined so that we could differentiate between 'let f x'
and 'let x'. They are not needed anymore however, as the parser
can remove the syntax sugar during parsing and convert 'let f x = e' into
'let f := fn x => e'.
The let definition would be recursive (meaning that the defined name
is bound in the definition body) if and only if the right side
of the assignment is a lambda expression.