Changes in lambdas and qualified names.
Lambdas were changed so that they can only have a single parameter.
All multi-parameter lambdas are desugared to unary function so that
we get automatic currying for free. Since optimizations will be
performed on different type of internal representation, this change
wouldn't really have a negative impact on performance of the
compiled code.

Extern constructor was renamed to External.

ScopedName was changed so that the prefix can only be a ModName instead
of a full module path.
The reason is that we should only allow the usage of explicitly imported
modules. If the user haven't imported module Foo\Bar, then he shouldn't
be able to use Foo\Bar\foo in the code.
In the current (unfinished) implementation of the parser a lot of effort
is spent on finding and including the imported files. Random qualified
names from modules that haven't been imported add another difficulty to
the already complex parser.

This means the following code will be invalid:
```
module FooBar

let foo := Foo\Bar\foo
```
But the following will be valid:
```
module FooBar
import Foo\Bar as Foo

let foo := Foo\foo
```