Separated Identifiers from Constructors.
Previously we used Var both for variables and constructors. I thought
it may be a good idea to separate them in the AST so that we won't
need to check if they're constructors or normal variables using
string comparisons.

Lambdas now may have T.Text instead of Identifier

Renamed ScopedVar into QualifiedVar and added QualifiedConstructor.