Better errors for operator redeclaration.
Previously the errors were reported after skipping the trailing whitespace.
This led to to weird error messages that would point at the next
statement instead of the one that caused the error. I fixed it
using "posFail" function which reports failures by changing the
current offset.
This solution is not ideal (I should take a look at the custom
error components).