Added unit tests for GetOpt module.
I've added several tests to check whether the command line arguments
are being parsed correctly. To do that I've split the module
into GetOpt and GetOpt.Internal so that the local functions are somewhat
hidden from the user but can be accessed by tests by importing the
GetOpt.Internal module. Due to the fact that parsing and effects of the
arguments are intertwined, it wasn't possible to test all existing flags.