Defined the intermediate AST type.
I've defined the type which describes a half-parsed module. This type
is produced by parsing a single file. These ASTs can be then chained
together by the module resolver so that they can create one normal AST.

It is unfortunate that this solution requires so much repetition but the
only way to avoid it would require a change in the normal AST definition.